Refactored export codes
