+++ espnet_onnx/export/asr/models/__init__.py
@@ -1,5 +1,5 @@
-from .ctc import CTC
-from .joint_network import JointNetwork
+from espnet_onnx.export.asr.models.ctc import CTC
+from espnet_onnx.export.asr.models.joint_network import JointNetwork
 
 # encoder
 from espnet2.asr.encoder.rnn_encoder import RNNEncoder as espnetRNNEncoder
@@ -8,20 +8,16 @@ from espnet2.asr.encoder.contextual_block_transformer_encoder import ContextualB
 from espnet2.asr.encoder.contextual_block_conformer_encoder import ContextualBlockConformerEncoder as espnetContextualConformer
 from espnet2.asr.encoder.transformer_encoder import TransformerEncoder as espnetTransformerEncoder
 from espnet2.asr.encoder.conformer_encoder import ConformerEncoder as espnetConformerEncoder
-from .encoders.rnn import RNNEncoder
-from .encoders.transformer import TransformerEncoder
-from .encoders.conformer import ConformerEncoder
-from .encoders.contextual_block_xformer import ContextualBlockXformerEncoder
+from espnet_onnx.export.asr.models.encoders.rnn import RNNEncoder
+from espnet_onnx.export.asr.models.encoders.transformer import TransformerEncoder
+from espnet_onnx.export.asr.models.encoders.conformer import ConformerEncoder
+from espnet_onnx.export.asr.models.encoders.contextual_block_xformer import ContextualBlockXformerEncoder
 
 # decoder
 from espnet2.asr.decoder.rnn_decoder import RNNDecoder as espnetRNNDecoder
-from espnet2.asr.decoder.transformer_decoder import TransformerDecoder as espnetTransformerDecoder
-from .decoders.rnn import (
-    RNNDecoder,
-    PreDecoder
-)
-from .decoders.xformer import XformerDecoder
-from .decoders.transducer import TransducerDecoder
+from espnet_onnx.export.asr.models.decoders.rnn import RNNDecoder
+from espnet_onnx.export.asr.models.decoders.xformer import XformerDecoder
+from espnet_onnx.export.asr.models.decoders.transducer import TransducerDecoder
 try:
     from espnet2.asr.transducer.transducer_decoder import TransducerDecoder as espnetTransducerDecoder
 except:
@@ -31,12 +27,12 @@ except:
 # lm
 from espnet2.lm.seq_rnn_lm import SequentialRNNLM as espnetSequentialRNNLM
 from espnet2.lm.transformer_lm import TransformerLM as espnetTransformerLM
-from .language_models.seq_rnn import SequentialRNNLM
-from .language_models.transformer import TransformerLM
+from espnet_onnx.export.asr.models.language_models.seq_rnn import SequentialRNNLM
+from espnet_onnx.export.asr.models.language_models.transformer import TransformerLM
 
 # frontend
 from espnet2.asr.frontend.s3prl import S3prlFrontend as espnetS3PRLModel
-from .frontends.s3prl import S3PRLModel
+from espnet_onnx.export.asr.models.frontends.s3prl import S3PRLModel
 
 
 def get_encoder(model, frontend, preencoder, export_config):
@@ -71,5 +67,3 @@ def get_frontend_models(model, export_config):
         return S3PRLModel(model, **export_config)
     else:
         return None
-    
-
