+++ espnet_onnx/export/asr/models/ctc.py
@@ -1,6 +1,5 @@
 import os
 
-import numpy as np
 import torch
 import torch.nn as nn
