+++ espnet_onnx/export/asr/models/decoder_layer.py
@@ -1,17 +1,7 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-# Copyright 2019 Shigeki Karita
-#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
-
 """Encoder self-attention layer definition."""
 
-from espnet_onnx.utils.torch_function import normalize
 import torch
-
 from torch import nn
-
-from espnet.nets.pytorch_backend.transformer.layer_norm import LayerNorm
 
 
 class OnnxDecoderLayer(nn.Module):
