+++ espnet_onnx/export/asr/models/frontends/s3prl.py
@@ -1,26 +1,12 @@
 import os
 
-import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from espnet.nets.pytorch_backend.transformer.subsampling import (
-    Conv2dSubsampling,
-    Conv2dSubsampling2,
-    Conv2dSubsampling6,
-    Conv2dSubsampling8
-)
-from espnet2.asr.frontend.default import DefaultFrontend
-from espnet2.layers.global_mvn import GlobalMVN
-from espnet2.layers.utterance_mvn import UtteranceMVN
-
-from espnet.nets.pytorch_backend.transformer.attention import MultiHeadedAttention
-
+from espnet_onnx.export.asr.models.encoder_layer import OnnxEncoderLayer
+from espnet_onnx.export.asr.models.multihead_att import OnnxMultiHeadedAttention
 from espnet_onnx.utils.torch_function import MakePadMask
-from ..language_models.embed import Embedding
-from ..encoder_layer import OnnxEncoderLayer
-from ..multihead_att import OnnxMultiHeadedAttention
 from espnet_onnx.utils.abs_model import AbsExportModel
