+++ espnet_onnx/export/asr/models/joint_network.py
@@ -1,6 +1,5 @@
 import os
 
-import numpy as np
 import torch
 import torch.nn as nn
