+++ espnet_onnx/export/layers/predecoder.py
@@ -1,11 +1,10 @@
-
+import os
 import torch
 import torch.nn as nn
 
 from espnet.nets.pytorch_backend.rnn.attentions import NoAtt
 
-from espnet_onnx.utils.function import make_pad_mask
-from espnet_onnx.export.layers.attention import get_attention, OnnxNoAtt, require_tanh
+from espnet_onnx.export.layers.attention import require_tanh
 from espnet_onnx.utils.abs_model import AbsExportModel
 
 
@@ -49,4 +48,3 @@ class PreDecoder(nn.Module, AbsExportModel):
         return {
             "model_path": os.path.join(path, f'{self.model_name}.onnx'),
         }
-
