+++ espnet_onnx/export/tts/models/tts_models/fastspeech2.py
@@ -1,10 +1,5 @@
+from typing import Optional
 
-from typing import (
-    Optional,
-    Tuple
-)
-
-import math
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
