+++ espnet_onnx/export/tts/models/tts_models/tacotron2.py
@@ -1,12 +1,7 @@
-
-from typing import (
-    Optional,
-    Tuple
-)
+from typing import Optional
 import six
 import os
 
-import math
 import numpy as np
 import torch
 import torch.nn as nn
@@ -17,17 +12,13 @@ from espnet.nets.pytorch_backend.rnn.attentions import (
     AttForwardTA
 )
 
-from espnet_onnx.export.asr.models.language_models.embed import Embedding
+from espnet_onnx.export.layers.predecoder import PreDecoder
+from espnet_onnx.export.layers.attention import get_attention
 from espnet_onnx.utils.torch_function import (
     MakePadMask,
     normalize
 )
 from espnet_onnx.utils.abs_model import AbsExportModel
-from espnet_onnx.export.layers.predecoder import PreDecoder
-from espnet_onnx.export.layers.attention import (
-    get_attention,
-    OnnxAttForward,
-)
 
 
 class OnnxEncoderLayer(nn.Module):
