+++ espnet_onnx/export/tts/models/vocoders/hifigan.py
@@ -1,10 +1,5 @@
+from typing import Optional
 
-from typing import (
-    Optional,
-    Tuple
-)
-
-import math
 import torch
 import torch.nn as nn
