+++ espnet_onnx/export/tts/models/vocoders/style_melgan.py
@@ -1,9 +1,3 @@
-
-from typing import (
-    Optional,
-    Tuple
-)
-
 import math
 import torch
 import torch.nn as nn
